feat: app cleanup
